requirements.txt
cachetools==5.3.0
charset-normalizer==3.1.0
click==8.1.3
Deprecated==1.2.13
Flask==2.2.3
Flask-Limiter==3.3.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
Flask-SSLify==0.1.5
gunicorn==20.1.0
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
limits==3.3.1
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.1
Pygments==2.15.0
pyngrok==5.2.1
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rich==13.3.4
six==1.16.0
SQLAlchemy==2.0.8
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wrapt==1.15.0
